--- src/topology/tests/index.js
@@ -1,9 +1,0 @@
-/* eslint-env mocha */
-
-'use strict'
-
-const topologySuite = require('./topology')
-
-module.exports = (test) => {
-  topologySuite(test)
-}
